Move deep link handling to App component
